добавлен schedule и удален student

+++ headteacher-service/src/main/java/com/example/headteacherservice/repository/GroupRepository.java
@@ -4,8 +4,11 @@ import com.example.headteacherservice.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 
     void deleteByGroupName(String groupName);
+
 }
